STY: renamed files for consistency
