+++ pyxrf/gui_module/tab_wd_plots_preview.py
