+++ pyxrf/gui_module/tab_wd_plots_rgb_maps.py
